refactor: Separated Commands in CLI
